requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
google-genai>=0.8.0
PyMuPDF>=1.24.9
requests>=2.31.0
pydantic>=2.7.0
matplotlib
seaborn
pandas>=2.2.0
yfinance>=0.2.54
scikit-learn>=1.5.0
imbalanced-learn>=0.12.0
